pyproject.toml
[project]
name = "personal-assistant"
version = "0.1.0"
description = "Modular AI assistant with atomic components"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # LLM
    "litellm>=1.40.0",
    
    # Web framework
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    
    # HTTP client
    "httpx>=0.27.0",
    
    # Data validation
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    
    # Config
    "pyyaml>=6.0.1",
    
    # Database
    "asyncpg>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.0",
    
    # Redis
    "redis>=5.0.0",
    
    # Vector DB
    "qdrant-client>=1.8.0",
    
    # S3
    "boto3>=1.34.0",
    
    # HTML parsing
    "beautifulsoup4>=4.12.0",
    "lxml>=5.1.0",
    
    # PDF parsing
    "pymupdf>=1.24.0",
    
    # Async utilities
    "aiofiles>=23.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.3.0",
    "mypy>=1.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
